fix labs: 3, 6, 7. Labs 8, 9

+++ 3822B1FI1/6_block_gemm_oneapi/mironov_arseniy/block_gemm_oneapi.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 
 std::vector<float> GemmBlockONEAPI(const std::vector<float> &a,
-                                   const std::vector<float> &b, int size,
+                                   const std::vector<float> &b, size_t size,
                                    sycl::device device) {
 
   sycl::queue gpu_queue(device);
